# openapi-v2.yaml
openapi: 3.1.0
info:
  title: eLearning Template Engine API
  description: |
    Backend API for the composable component-based eLearning authoring platform.
    Supports 89+ component types, per-interaction audio, page completion tracking,
    weighted scoring, full theming (layout + color system), and SCORM 1.2/2004 export.
  version: 2.0.0
  contact:
    name: eLearning Platform Team
servers:
  - url: /api/v1
    description: API v1 namespace
...
paths:
...
  /courses/{courseId}/pages/{pageId}/components:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PageId"
    get:
      tags: [Components]
      operationId: listPageComponents
      summary: List components on a page (ordered)
      responses:
        "200":
          description: Ordered list of components
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComponentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Components]
      operationId: addComponent
      summary: Add a component to a page
      description: |
        Component `data` is validated against the JSON Schema from the Component Type Registry
        for the given `componentType`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentCreateRequest"
      responses:
        "201":
          description: Component created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /courses/{courseId}/pages/{pageId}/components/{componentId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PageId"
      - $ref: "#/components/parameters/ComponentId"
    get:
      tags: [Components]
      operationId: getComponent
      summary: Get a single component
      responses:
        "200":
          description: Component detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Components]
      operationId: updateComponent
      summary: Update a component
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentUpdateRequest"
      responses:
        "200":
          description: Updated component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Components]
      operationId: deleteComponent
      summary: Remove a component from a page
      responses:
        "204":
          description: Component deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Component Registry ──────────────────────
  /components:
    get:
      tags: [ComponentRegistry]
      operationId: listComponentTypes
      summary: List all component types
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category ID
        - name: scoringEnabled
          in: query
          schema:
            type: boolean
          description: Filter scorable types only
        - name: isActive
          in: query
          schema:
            type: boolean
            default: true
          description: Include inactive types
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated component type list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentTypeListResponse"
...
  /components/categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [ComponentRegistry]
      operationId: listCategoryComponents
      summary: List component types in a category
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Component types in this category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentTypeListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /components/search:
    get:
      tags: [ComponentRegistry]
      operationId: searchComponentTypes
      summary: Search component types by query/tags
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query (matches displayName, description, tags)
        - name: category
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
          description: Comma-separated tags
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentTypeListResponse"
  /components/{typeId}:
    parameters:
      - name: typeId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [ComponentRegistry]
      operationId: getComponentType
      summary: Get a component type definition
      description: Returns full definition including JSON Schema, default data, scoring rules
      responses:
        "200":
          description: Component type definition
          headers:
            ETag:
              schema:
                type: string
              description: Entity tag for conditional requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentTypeDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Themes ──────────────────────────────────
...
components:
...
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_-]+$"
      description: Course identifier
    PageId:
      name: pageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Page UUID
    ComponentId:
      name: componentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Component UUID
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200

  # ── Responses ───────────────────────────────
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ── Schemas ─────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ValidationErrorResponse:
      type: object
      required: [detail, errors]
      properties:
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    AudioItem:
      type: object
      required: [audioUrl, triggerOn]
      properties:
        audioId:
          type: string
          format: uuid
          readOnly: true
        audioUrl:
          type: string
          description: URL to audio file from media upload
        triggerOn:
          type: string
          enum: [load, click, interaction]
        targetInteractionId:
          type: string
          nullable: true
          description: "Interaction point ID (e.g. 'tab-2', 'accordion-panel-1')"
        autoplay:
          type: boolean
          default: false
        requiredForCompletion:
          type: boolean
          default: false
        duration:
          type: number
          description: Duration in seconds
        label:
          type: string
          nullable: true
          description: Display name for narration list
        transcript:
          type: string
          nullable: true
          description: Text transcript for accessibility
    AudioConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        audioItems:
          type: array
          items:
            $ref: "#/components/schemas/AudioItem"
          default: []
...
    CompletionCriteria:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [view, interact, audio, score, custom]
        threshold:
          type: number
          minimum: 0
          maximum: 100
          nullable: true
          description: Score threshold (for type=score)
        requiredInteractions:
          type: array
          items:
            type: string
          nullable: true
          description: Interaction IDs that must be performed
        requiredAudioIds:
          type: array
          items:
            type: string
          nullable: true
          description: Audio IDs that must be listened to

    # ── Component Styling ───────────────
    ComponentStyling:
      type: object
      properties:
        themeOverrides:
          $ref: "#/components/schemas/ThemeOverrides"
        layoutPosition:
          type: string
          nullable: true
          description: Grid area name from page layout

    # ── Component ───────────────────────
    ComponentCreateRequest:
      type: object
      required: [componentType, data]
      properties:
        componentType:
          type: string
          description: Must match a typeId from the Component Type Registry
        data:
          type: object
          description: Component-specific data validated against registry schema
        audioConfig:
          $ref: "#/components/schemas/AudioConfig"
        completionCriteria:
          $ref: "#/components/schemas/CompletionCriteria"
        styling:
          $ref: "#/components/schemas/ComponentStyling"
    ComponentUpdateRequest:
      type: object
      properties:
        data:
          type: object
        audioConfig:
          $ref: "#/components/schemas/AudioConfig"
        completionCriteria:
          $ref: "#/components/schemas/CompletionCriteria"
        styling:
          $ref: "#/components/schemas/ComponentStyling"
    ComponentResponse:
      type: object
      properties:
        componentId:
          type: string
          format: uuid
        componentType:
          type: string
        order:
          type: integer
        data:
          type: object
        audioConfig:
          $ref: "#/components/schemas/AudioConfig"
        completionCriteria:
          $ref: "#/components/schemas/CompletionCriteria"
        styling:
          $ref: "#/components/schemas/ComponentStyling"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Page Completion ─────────────────
...
    AudioSupportConfig:
      type: object
      properties:
        perComponent:
          type: boolean
          description: Can have component-level audio
        perInteraction:
          type: boolean
          description: Can have audio per tab/panel/item
        interactionPoints:
          type: array
          items:
            type: string
          nullable: true
          description: "Names of interaction points (e.g. ['tab', 'panel'])"
    ComponentTypeSummary:
      type: object
      properties:
        typeId:
          type: string
        category:
          type: string
        displayName:
          type: string
        description:
          type: string
        icon:
          type: string
        thumbnail:
          type: string
          nullable: true
        completionCapabilities:
          type: array
          items:
            type: string
            enum: [view, interact, audio, score]
        scoringEnabled:
          type: boolean
        audioSupport:
          $ref: "#/components/schemas/AudioSupportConfig"
        tags:
          type: array
          items:
            type: string
        estimatedDuration:
          type: number
          nullable: true
    ComponentTypeDetail:
      allOf:
        - $ref: "#/components/schemas/ComponentTypeSummary"
        - type: object
          properties:
            schemaVersion:
              type: string
              description: Schema version for cache invalidation
            etag:
              type: string
              description: ETag for conditional requests
            schema:
              type: object
              description: JSON Schema for validating component data
            defaultData:
              type: object
              description: Default data when component is first added
            defaultCompletionType:
              type: string
            maxScore:
              type: number
              nullable: true
            scoringRules:
              type: object
              nullable: true
              description: Per-quiz-type scoring configuration
            sortOrder:
              type: integer
            isActive:
              type: boolean
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    ComponentTypeListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ComponentTypeSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        schemaVersion:
          type: string
          description: Schema version for cache invalidation
...
    ThemeOverrides:
      type: object
      description: Partial theme — any subset of properties can be overridden
      properties:
        colors:
          type: object
          additionalProperties:
            type: string
        typography:
          type: object
          additionalProperties: true
        componentStyles:
          type: object
          additionalProperties: true
...